Redesign UX (#3)
* Added Ribbon
* Show Icon on tabs
* Zoom in/out for editors
* Show validation errors
* Use MainWindow on DialogService
* Show/Hide System Properties on document
* Add External component list in about box

+++ src/DocumentDbExplorer/ViewModel/DatabaseNodes/CollectionNodeViewModel.cs
@@ -9,7 +9,8 @@ using Microsoft.Azure.Documents;
 
 namespace DocumentDbExplorer.ViewModel
 {
-    public class CollectionNodeViewModel : TreeViewItemViewModel
+
+    public class CollectionNodeViewModel : TreeViewItemViewModel, IHaveCollectionNodeViewModel
     {
         private readonly IDialogService _dialogService;
         private readonly IDocumentDbService _dbService;
@@ -141,12 +142,17 @@ namespace DocumentDbExplorer.ViewModel
                             await _dialogService.ShowMessage("Are you sure you want to delete this collection?", "Delete", null, null,
                                 async confirm =>
                                 {
-                                    await _dbService.DeleteCollection(Parent.Parent.Connection, Collection);
-                                    await DispatcherHelper.RunAsync(() => Parent.Children.Remove(this));
+                                    if (confirm)
+                                    {
+                                        await _dbService.DeleteCollection(Parent.Parent.Connection, Collection);
+                                        await DispatcherHelper.RunAsync(() => Parent.Children.Remove(this));
+                                    }
                                 });
                         }
                         ));
             }
         }
+
+        public CollectionNodeViewModel CollectionNode => this;
     }
 }
